+++ src/Regressors/Adaline.php
@@ -13,7 +13,7 @@ use Rubix\ML\EstimatorType;
 use Rubix\ML\Helpers\Params;
 use Rubix\ML\Datasets\Dataset;
 use Rubix\ML\Traits\LoggerAware;
-use Rubix\ML\NeuralNet\FeedForward;
+use Rubix\ML\NeuralNet\Network;
 use Rubix\ML\NeuralNet\Layers\Dense;
 use Rubix\ML\Traits\AutotrackRevisions;
 use Rubix\ML\NeuralNet\Optimizers\Adam;
@@ -108,9 +108,9 @@ class Adaline implements Estimator, Learner, Online, RanksFeatures, Verbose, Per
     /**
      * The underlying neural network instance.
      *
-     * @var \Rubix\ML\NeuralNet\FeedForward|null
+     * @var \Rubix\ML\NeuralNet\Network|null
      */
-    protected ?\Rubix\ML\NeuralNet\FeedForward $network = null;
+    protected ?\Rubix\ML\NeuralNet\Network $network = null;
 
     /**
      * The loss at each epoch from the last training session.
@@ -260,9 +260,9 @@ class Adaline implements Estimator, Learner, Online, RanksFeatures, Verbose, Per
     /**
      * Return the underlying neural network instance or null if not trained.
      *
-     * @return \Rubix\ML\NeuralNet\FeedForward|null
+     * @return \Rubix\ML\NeuralNet\Network|null
      */
-    public function network() : ?FeedForward
+    public function network() : ?Network
     {
         return $this->network;
     }
@@ -276,7 +276,7 @@ class Adaline implements Estimator, Learner, Online, RanksFeatures, Verbose, Per
     {
         DatasetIsNotEmpty::with($dataset)->check();
 
-        $this->network = new FeedForward(
+        $this->network = new Network(
             new Placeholder1D($dataset->numFeatures()),
             [new Dense(1, $this->l2Penalty, true, new Xavier2())],
             new Continuous($this->costFn),